[hw-4] refactor: rewrite models and services
