+++ homework_4/src/models/user.ts
@@ -1,77 +1,31 @@
-import { UserDB } from './userDBSchema';
-import { 
-  EditablePropsOfUser,
-  UpdatingPropsOfUser, 
-  UserModelInterface
-} from '../interfaces/user';
+import { dbConnection } from '../data-access/dbInit';
+import { DataTypes } from 'sequelize';
 
-const { Op } = require("sequelize");
-
-export class UserModel implements UserModelInterface {
-  async addUser(newUserData: EditablePropsOfUser) {
-    const { id } = await UserDB.create(newUserData);
-
-    return String(id);
+const UserDB = dbConnection.define('users', {
+  id: {
+    primaryKey: true,
+    autoIncrement: true,
+    type: DataTypes.INTEGER,        
+  },
+  isDeleted: {
+    type: DataTypes.BOOLEAN,
+    defaultValue: false,
+  },
+  login: {
+    type: DataTypes.STRING,
+    allowNull: false
+  },
+  password: {
+    type: DataTypes.STRING,
+    allowNull: false,
+  },
+  age: {
+    type: DataTypes.INTEGER,
+    allowNull: false,
   }
+});
 
-  async getUser(userId: string) {
-    const user = await UserDB.findOne({
-      where: {
-        id: userId,
-        isDeleted: false,
-      },
-      raw: true,
-    });
-  
-    let result = null;
+UserDB.sync()
+  .catch((err: Error)=> console.log(err));
 
-    if (user) {
-      result = {
-        id: user.id,
-        login: user.login,
-        password: user.password,
-        age: user.age,
-      };
-    }
-    return result;
-  }
-
-  async updateUser(userId: string, updatedUserData: UpdatingPropsOfUser) {
-    await UserDB.update(updatedUserData, {
-      where: {
-        isDeleted: false,
-        id: userId
-      }
-    });
-  }
-
-  async removeUser(userId: string) {
-    await UserDB.update({ isDeleted: true }, {
-      where: {
-        id: userId,
-        isDeleted: false,
-      }
-    });    
-  }
-
-  async getAutoSuggestUsers(loginSubstring: string, limit: number) {
-    const dbFoundUsers = await UserDB.findAll({ limit, order: [['login', 'DESC']] }, {
-      where: {
-        isDeleted: false,
-        login: {
-          [Op.like]: loginSubstring,
-        }
-      }
-    });
-    const result = dbFoundUsers.map((dbUser: any) => {
-      return {
-        id: dbUser.id,
-        login: dbUser.login,
-        password: dbUser.password,
-        age: dbUser.age,
-      }
-    });
-    
-    return result;
-  }
-}
+export { UserDB };
