+++ homework_4/src/services/user.ts
@@ -1,29 +1,79 @@
+import { Op } from 'sequelize';
 import {
   EditablePropsOfUser,
   UpdatingPropsOfUser,
-  UserModelInterface,
+  UserServiceInterface,
 } from '../interfaces/user';
 
-export class UserService {
-  constructor(private userModel: UserModelInterface) {
+export class UserService implements UserServiceInterface {
+  constructor(private userModel: any) {
     this.userModel = userModel;
   }
 
-  createNewUser = async (newUserData: EditablePropsOfUser) => {
-    return await this.userModel.addUser(newUserData);
+  create = async (newUserData: EditablePropsOfUser) => {
+    const { id } = await this.userModel.create(newUserData);
+
+    return String(id);
   };
 
-  updateUser = async (userId: string, updatingUserData: UpdatingPropsOfUser) => {
-    const currentUserData = await this.userModel.getUser(userId);
-    const updatedUserData = Object.assign({}, currentUserData, updatingUserData);
-    return await this.userModel.updateUser(userId, updatedUserData);
+  update = async (userId: string, updatingUserData: UpdatingPropsOfUser) => {
+    await this.userModel.update(updatingUserData, {
+      where: {
+        isDeleted: false,
+        id: userId
+      }
+    });
   };
 
-  getUser = async (userId: string) => await this.userModel.getUser(userId);
+  get = async (userId: string) => {
+    const user = await this.userModel.findOne({
+      where: {
+        id: userId,
+        isDeleted: false,
+      },
+      raw: true,
+    });
+  
+    let result = null;
 
-  removeUser = async (userId: string) => await this.userModel.removeUser(userId);
+    if (user) {
+      result = {
+        id: user.id,
+        login: user.login,
+        password: user.password,
+        age: user.age,
+      };
+    }
+    return result;
+  }
+
+  remove = async (userId: string) => {
+    await this.userModel.update({ isDeleted: true }, {
+      where: {
+        id: userId,
+        isDeleted: false,
+      }
+    });
+  }
 
   getAutoSuggestUsers = async (loginSubstring: string, limit: number) => {
-    return await this.userModel.getAutoSuggestUsers(loginSubstring, limit);
+    const dbFoundUsers = await this.userModel.findAll({ limit, order: [['login', 'DESC']] }, {
+      where: {
+        isDeleted: false,
+        login: {
+          [Op.like]: loginSubstring,
+        }
+      }
+    });
+    const result = dbFoundUsers.map((dbUser: any) => {
+      return {
+        id: dbUser.id,
+        login: dbUser.login,
+        password: dbUser.password,
+        age: dbUser.age,
+      }
+    });
+    
+    return result;
   }
 }
