+++ homework_4/src/services/userGroup.ts
@@ -1,18 +1,31 @@
+import { Op } from 'sequelize';
 import {
-  UserGroupModelInterface,
+  UserGroupServiceInterface,
   UsersToGroupParams
 } from '../interfaces/userGroup';
 
-export class UserGroupService {
-  constructor(private userGroupModel: UserGroupModelInterface) {
+export class UserGroupService implements UserGroupServiceInterface {
+  constructor(private userGroupModel: any) {
     this.userGroupModel = userGroupModel;
   }
 
   addUsersToGroup = async (data: UsersToGroupParams) => {
-    return await this.userGroupModel.addUsersToGroup(data);
+    const { groupId, usersId } = data;
+    usersId.forEach(userId => {
+      this.userGroupModel.create({ groupId, userId: userId });
+    });
   };
 
   deleteUsersFromGroup = async (data: UsersToGroupParams) => {
-    return await this.userGroupModel.deleteUsersFromGroup(data);
+    const { groupId, usersId } = data;
+
+    await this.userGroupModel.destroy({
+      where: {
+        groupId: groupId,
+        userId: {
+          [Op.or]: usersId
+        }
+      }
+    });
   };
 }
